Update compressor and decompressor benchmarks with Google benchmark
